refactor: Database initialization strategy (#20)
* optimizations

* header prep strategy

* block build strategy

* db init strategy

* nit
