+++ lib/src/lib.rs
@@ -21,9 +21,10 @@ pub mod block_builder;
 pub mod consts;
 pub mod execution;
 pub mod finalization;
+pub mod initialization;
+pub mod input;
 pub mod mem_db;
 pub mod preparation;
-pub mod validation;
 
 /// call forget only if running inside the guest
 pub fn guest_mem_forget<T>(_t: T) {
