+++ testing/ef-tests/testguest/src/main.rs
@@ -16,11 +16,10 @@
 
 use risc0_zkvm::guest::env;
 use zeth_lib::{
-    block_builder::BlockBuilder, consts::ChainSpec, execution::EthTxExecStrategy, mem_db::MemDb,
-    validation::Input,
+    block_builder::BlockBuilder, consts::ChainSpec, execution::EthTxExecStrategy,
+    finalization::BuildFromMemDbStrategy, initialization::MemDbInitStrategy, input::Input,
+    mem_db::MemDb, preparation::EthHeaderPrepStrategy,
 };
-use zeth_lib::finalization::BuildFromMemDbStrategy;
-use zeth_lib::preparation::EthHeaderPrepStrategy;
 
 risc0_zkvm::guest::entry!(main);
 
@@ -31,7 +30,7 @@ pub fn main() {
     let input: Input = env::read();
     // Build the resulting block
     let output = BlockBuilder::<MemDb>::new(&chain_spec, input)
-        .initialize_db()
+        .initialize_database::<MemDbInitStrategy>()
         .expect("Failed to create in-memory evm storage")
         .prepare_header::<EthHeaderPrepStrategy>()
         .expect("Failed to create the initial block header fields")
